Generalize the pipeline of main.py
The call to function `navigate_gdrive`, which is specific to
"navigate" commnad, is handled through a wrapper via navigate
subparser. This wrapper follows a unify pattern of signature, taking
`args` and `configs` instead of args specific to each wrapped
function.

To this end, the configuration is now represented as a
`SimpleNamespace` instead of a `dict`.

Further simplifications invlove removal of the cli arguments for
navigate subparser. Now `remote` is read from config file, and
`start_path` is left to the default value `""` of the
`navigate_gdrive`.

Signed-off-by: Saeed Gholami Shahbandi <saeed.gh.sh@gmail.com>