Clean up test_rclone_wrapper.py
Some tests were checking the exact details of the raised error
messages which was not a good idea. There is no need for testing the
specifics of implementation, just the behavior is enough.

Some tests implemented "try/except/else" block that was unnecessary
because `pytest` has built-in support for asserting exceptions with
`pytest.raises`.

Signed-off-by: Saeed Gholami Shahbandi <saeed.gh.sh@gmail.com>